requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
Cython==3.1.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.40.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
frozenlist==1.6.0
graphql-core==3.2.6
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jwt==1.3.1
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.4.4
numpy==2.2.6
openai==1.82.0
packaging==25.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-engineio==4.12.1
python-jose==3.3.0
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
simple-websocket==1.1.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.20.0
